Некоторые правки HW-1 #1

+++ project/include/utils.h
@@ -1,13 +1,11 @@
-#include <stddef.h>
-#include <stdio.h>
-#include <stdlib.h>
 #ifndef PROJECT_INCLUDE_UTILS_H_
 #define PROJECT_INCLUDE_UTILS_H_
 
+#include <stddef.h>
 
-size_t timer_from(int from);
-// TODO(stitaevskiy): Implement `power of` function
-int custom_pow(int base, int power);
+size_t timer_from(long int from);
+
+long int custom_pow(long int base, long int power);
 
 #endif  // PROJECT_INCLUDE_UTILS_H_
